[#4] Refactor PatientRecord model to change the data type of `fasting_blood_sugar` from integer to float. This change is necessary to allow for more precise recording of fasting blood sugar levels.

+++ db/migrate/20231130153552_create_patient_records.rb
@@ -8,7 +8,7 @@ class CreatePatientRecords < ActiveRecord::Migration[7.1]
       t.integer :chest_pain_type, null: false
       t.float :resting_blood_pressure
       t.float :serum_cholesterol
-      t.integer :fasting_blood_sugar
+      t.float :fasting_blood_sugar
       t.integer :resting_ecg_results, null: false
       t.integer :max_heart_rate_achieved
       t.boolean :exercise_induced_angina, default: false
